Hammad and Chris
Working on changing the event object to add rounds to the fetch for display on the public user. Currently the rounds.map function in the event specific page.tsx is having an error but should be a easy fix. Then will be displayed along the divisions for easy retrieval of results.